+++ src/job/jigdo-io.hh
@@ -113,7 +113,7 @@ private:
   virtual void job_failed(string* message);
   virtual void job_message(string* message);
   virtual void dataSource_dataSize(uint64 n);
-  virtual void dataSource_data(const byte* data, size_t size,
+  virtual void dataSource_data(const byte* data, unsigned size,
                                uint64 currentSize);
 
   // Virtual methods from Gunzip::IO
